% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{get_weights}
\alias{get_weights}
\title{Extract sample weights}
\usage{
get_weights(object)
}
\arguments{
\item{object}{An objects of class 'balancing_weights'.}
}
\value{
Empirically normalized sample weights.
}
\description{
A generic function to extract sample weights from objects of class 'balancing_weights'
}
\examples{
data(demo)
output1 <- balancing.weights(S, Z, X, method = "IC", naturalGroupProp)
get_weights(output1)
}
