% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{mean_diff}
\alias{mean_diff}
\title{Extract causal estimates (mean differences)}
\usage{
mean_diff(object)
}
\arguments{
\item{object}{An objects of class 'causal_estimates'.}
}
\value{
The mean differences between two groups
}
\description{
A generic function to extract mean differences for objects of class 'causal_estimates'.
}
\examples{
data(demo)
output2 <- causal.estimate(S, Z, X, Y, B = 5, method = "IC", naturalGroupProp)
mean_diff(output2)

}
