% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{percentESS}
\alias{percentESS}
\title{Extract percentage sample ESS}
\usage{
percentESS(object)
}
\arguments{
\item{object}{An objects of class 'balancing_weights' or 'causal_estimates'.}
}
\value{
Percentage sample effective sample size (ESS) for the pseudo-population.
}
\description{
A generic function to extract percentage sample ESS for different object classes
}
\examples{
data(demo)
output1 <- balancing.weights(S, Z, X, method = "IC", naturalGroupProp)
percentESS(output1)

output2 <- causal.estimate(S, Z, X, Y, B = 5, method = "IC", naturalGroupProp)
percentESS(output2)

}
