% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.causal_estimates}
\alias{plot.causal_estimates}
\title{Plot method for objects of class 'causal_estimates'}
\usage{
\method{plot}{causal_estimates}(x, ...)
}
\arguments{
\item{x}{An object of class 'causal_estimates'.}

\item{...}{Additional arguments including:
\describe{
\item{y_limit}{The y-axis range. Default is \code{c(0,50)}.}
\item{color}{The boxplot color. Default is "red".}
}}
}
\value{
A boxplot of percent sample ESS for a specific weighting method (\code{FLEXOR}, \code{IC}, or \code{IGO})
}
\description{
Plot method for objects of class 'causal_estimates'
}
