% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.balancing_weights}
\alias{print.balancing_weights}
\title{Print method for objects of class 'balancing_weights'}
\usage{
\method{print}{balancing_weights}(x, ...)
}
\arguments{
\item{x}{An object of class 'balancing_weights'}

\item{...}{Additional arguments affecting the printed results (so far no additional arguments are needed, so leave blank).}
}
\value{
Print values of the 'balancing_weights' object, including:

\describe{
\item{Weight length}{The total number of weights.}
\item{percentESS}{Percentage sample effective sample size (ESS) for the pseudo-population.}
}
}
\description{
Print method for objects of class 'balancing_weights'
}
\examples{
data(demo)
output1 <- balancing.weights(S, Z, X, method = "IC", naturalGroupProp)
print(output1)

}
