% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.causal_estimates}
\alias{print.causal_estimates}
\title{Print method for objects of class 'causal_estimates'}
\usage{
\method{print}{causal_estimates}(x, ...)
}
\arguments{
\item{x}{An object of class 'causal_estimates'}

\item{...}{Additional arguments affecting the printed results (so far no additional arguments are needed, so leave blank).}
}
\value{
Print values of the 'causal_estimates' object, including:

\describe{
\item{Percentage sample ESS}{Percentage sample effective sample size (ESS) for the pseudo-population.}
\item{Mean differences}{The mean differences between two groups}
\item{Sigma ratios}{The ratios of standard deviations between two groups}
}
}
\description{
Print method for objects of class 'causal_estimates'
}
\examples{
data(demo)
output2 <- causal.estimate(S, Z, X, Y, B = 5, method = "IC", naturalGroupProp)
print(output2)

}
