% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor.R
\name{sigma_ratio}
\alias{sigma_ratio}
\title{Extract sigma ratios}
\usage{
sigma_ratio(object)
}
\arguments{
\item{object}{An objects of class 'causal_estimates'.}
}
\value{
The ratios of standard deviations between two groups
}
\description{
A generic function to extract the ratios of standard deviations for objects of class 'causal_estimates'.
}
\examples{
data(demo)
output2 <- causal.estimate(S, Z, X, Y, B = 5, method = "IC", naturalGroupProp)
mean_diff(output2)

}
