% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.balancing_weights}
\alias{summary.balancing_weights}
\title{Summary method for objects of class 'balancing_weights'}
\usage{
\method{summary}{balancing_weights}(object, ...)
}
\arguments{
\item{object}{An object of class 'balancing_weights'}

\item{...}{Additional arguments affecting the summary produced (so far no additional arguments are needed, so leave blank).}
}
\value{
Printed summary of the 'balancing_weights' object, including:

\describe{
\item{Weight length}{The total number of weights.}
\item{Weight distribution}{Statistical summary of weight values.}
\item{Percentage sample ESS}{Percentage sample effective sample size (ESS) for the pseudo-population.}
}
}
\description{
Summary method for objects of class 'balancing_weights'
}
\examples{
data(demo)
output1 <- balancing.weights(S, Z, X, method = "IC", naturalGroupProp)
summary(output1)

}
