% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.causal_estimates}
\alias{summary.causal_estimates}
\title{Summary method for objects of class 'causal_estimates'}
\usage{
\method{summary}{causal_estimates}(object, ...)
}
\arguments{
\item{object}{An object of class 'causal_estimates'}

\item{...}{Additional arguments affecting the summary produced (so far no additional arguments are needed, so leave blank).}
}
\value{
Printed summary of the 'causal_estimates' object, including:

\describe{
\item{Percentage sample ESS}{Percentage sample effective sample size (ESS) for the pseudo-population.}
\item{Mean differences with 95\% CI}{The mean differences between two groups with their corresponding 95\% confidence intervals.}
\item{Sigma ratios with 95\% CI}{The ratios of standard deviations between two groups with their corresponding 95\% confidence intervals.}
}
}
\description{
Summary method for objects of class 'causal_estimates'
}
\examples{
data(demo)
set.seed(1)
output2 <- causal.estimate(S, Z, X, Y, B = 5, method = "IC", naturalGroupProp)
summary(output2)


}
